Ability of Marketing team to see landing page of "Artisans Database" tab in Android application(Partially completed)

+++ app/src/main/java/com/adrosonic/craftexchangemarketing/repository/data/response/admin/userDatabase/DatabaseResponse.kt
@@ -7,7 +7,7 @@ data class DatabaseResponse (
     var errorCode: Long
 )
 data class User(
-    var id :Int,
+    var id :Long,
     var weaverId : String,
     var rating : Float,
     var status : Int,
@@ -18,5 +18,5 @@ data class User(
     var lastName : String,
     var mobile : String,
     var dateAdded :String
-
 )
+
